requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pandas
google-generativeai
pydantic-settings